requirements.txt
accelerate==1.6.0
fairseq==0.12.2
flash_attn==2.7.1.post4
huggingface_hub==0.30.1
joblib==1.4.2
librosa==0.11.0
mel_cepstral_distance==0.0.4
mistral_common==1.8.1
omegaconf==2.0.5
openai_whisper==20240930
peft==0.15.1
pesq==0.0.4
pystoi==0.4.1
regex==2024.11.6
soundfile==0.13.1
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
transformers==4.53.0
transformers_stream_generator==0.0.5
